Clean up auth command layer: state helper + error-code normalization
#1 (state): add auth-state `beginAuthTransition()` — sets 'loading' and returns
a revert fn — so auth-commands stops juggling raw getAuthState/setState/
toStableAuthState. signIn/signOut now express intent, not state mechanics.

#2 (error normalization): move the Firebase `auth/...` → generic code map into a
pure leaf `shared/auth-error-codes.js` (no firebase import, so UI can use it;
adapter re-exports it). auth-commands, onetap, password-auth, and
UsernamePasswordForm now branch on normalized codes instead of raw strings,
closing the vendor-boundary exceptions ARCHITECTURE.md tracked.

Alerts stay inline in the imperative sign-in commands (legitimate user feedback);
only the state mechanics and code-branching were extracted. Existing password-auth
and form tests pass unchanged (the pure leaf runs in them).
